refactor: 인덱스 검증 유틸 클래스 패키지 이동, 테스트 작성
