+++ src/main/java/laddergame/domain/ladder/line/LineWidth.java
@@ -1,6 +1,6 @@
 package laddergame.domain.ladder.line;
 
-import laddergame.domain.util.ExceptionMessageFormatter;
+import laddergame.util.ExceptionMessageFormatter;
 
 public class LineWidth {
